// src/app/core/models/supplier.model.ts
export class SupplierDto {
    companyName: string;
    commercialLicenseNo: string;
    supplierClassification: string;
    supplierClassificationId: number;
    firstName: string;
    middleName: string;
    lastName: string;
    roleName: string;
    role: number;
    mobile: string;
    email: string;
    companyId: number;
    companyNameNative: string;
    parentCompanyName: string;
    supplierTypeId: number;
    organizationWebsite: string;
    commercialLicenseId: number;
    establishmentDate: string;
    issuedDate: string;
    issuedBy: string;
    supplierType: string;
    typeOfOwnership: string;
    typeOfOwnershipId: number;
    location: string;
    expiryDate: string;
    managerCommLicId: number;
    managerFirstName: string;
    managerMiddleName: string;
    managerLastName: string;
    managerRole: number;
    managerJobTitle: string
  }